requirements.txt
tornado
pylibmc
motor
toro
schematics
psycopg2
queries
pyzmq
arrow
phonenumbers
requests
pycurl
nose
ujson